Fix a number of compiler warnings and CodeQL findings
```
clock.cc: In function ‘int main(int, char**)’:
clock.cc:196:14: warning: ignoring return value of ‘ssize_t write(int, const void*, size_t)’, declared with attribute warn_unused_result [-Wunused-result]
  196 |   (void)write(STDOUT_FILENO, "\n", 1);  // Create a fresh new line after ^C on screen
      |         ~~~~~^~~~~~~~~~~~~~~~~~~~~~~~

rpi-rgb-led-matrix/lib/framebuffer.cc:299:
Multiplication result may overflow 'int' before it is converted to 'unsigned long'.
This rule finds code that converts the result of an integer multiplication to a larger type. Since the conversion applies after the multiplication, arithmetic overflow may still occur.

rpi-rgb-led-matrix/lib/framebuffer.cc:845:
Comparison between  of type uint8_t and  of wider type const int.
In a loop condition, comparison of a value of a narrow type with a value of a wide type may result in unexpected behavior if the wider value is sufficiently large (or small). This is because the narrower value may overflow. This can lead to an infinite loop.

examples-api-use/demo-main.cc:270
Multiplication result may overflow 'int' before it is converted to 'size_t'.
This rule finds code that converts the result of an integer multiplication to a larger type. Since the conversion applies after the multiplication, arithmetic overflow may still occur.

examples-api-use/ledcat.cc:43
Multiplication result may overflow 'int' before it is converted to 'ssize_t'.
This rule finds code that converts the result of an integer multiplication to a larger type. Since the conversion applies after the multiplication, arithmetic overflow may still occur.
```

+++ examples-api-use/ledcat.cc
@@ -40,7 +40,7 @@ int main(int argc, char *argv[]) {
   signal(SIGTERM, InterruptHandler);
   signal(SIGINT, InterruptHandler);
 
-  ssize_t frame_size = canvas->width() * canvas->height() * 3;
+  ssize_t frame_size = (ssize_t)canvas->width() * (ssize_t)canvas->height() * 3;
   uint8_t buf[frame_size];
 
   while (1) {
